chore(agents): add knip TS dead-code gate; move tests to test/
Introduce knip as the TypeScript counterpart of the aep-api Go dead-code
gate, scoped to @aep/agents + @aep/playground (playground counts as a
consumer, so agents exports it imports stay live).

- Gate: `make deadcode-ts-check` (report: `make deadcode-ts`) runs knip
  --production so *.test.ts never count as consumers (the Go "no -test").
  CI step added; config + rationale live in knip.jsonc. A `@knipkeep
  <reason>` JSDoc tag is the `//deadcode:keep` analogue — applied to the
  unwired HITL `askQuestionTool`.
- Close a blind spot: the @aep/agents `exports` map globbed *.test.ts in as
  public entry points, hiding test-only dead code. Tightened 3 wildcards to
  9 explicit consumed subpaths.
- Reorganize tests: all 18 *.test.ts + the 2 shared fixtures move src/** to
  a flat test/ dir (matching @aep/agent-stream and @aep/playground); src/
  ships zero test code. SEED_FILES relocated out of prompt.ts to
  test/seed-files.ts. mock-model stays in src/shared (published
  test-support via exports).

Verified: knip gate green; agents 147 + playground 32 tests pass;
typecheck, lint, license-check clean.
