+++ services/agents/src/agents/main/prompt.ts
@@ -122,88 +122,6 @@ export function buildTaskPlanInstructions(skills?: SkillSource): string {
   return taskPlanInstructions + buildSkillCatalog(skills);
 }
 
-/**
- * The starting spec bundle the demo mutates. Mirrors the hello-api example in
- * design.md: free-form prose, markdown-with-frontmatter, and indentation-
- * sensitive OpenAPI YAML — the three shapes the tools must handle.
- */
-export const SEED_FILES: Record<string, string> = {
-  "specs/requirements/requirements.md": `# Overview
-
-A simple API that responds with "Hello, World!" when called.
-
-# Personas
-
-- Developer — calls the API to get a hello world response.
-
-# Features
-
-- A developer sends a request to the API.
-- The API responds with "Hello, World!" in the response body.
-- The response is in JSON format with a message field.
-- The API is accessible via a single endpoint.
-- Requests work without requiring any parameters or authentication.
-`,
-
-  "specs/design/design.md": `A simple public API service that responds with "Hello, World!" in JSON format. Built as a single Go service exposing one endpoint, requiring no authentication.
-`,
-
-  "specs/design/components/hello-api/design.json": `{
-  "name": "hello-api",
-  "type": "service",
-  "version": "0.1.0",
-  "language": "Go",
-  "buildpack": "docker",
-  "appPath": "hello-api",
-  "entrypoint": "deployment/service",
-  "exposure": "internet",
-  "dependencies": [],
-  "skillsApplied": ["go", "api-management"],
-  "description": "A simple public Go HTTP service (port 9090, net/http) that returns a hello-world JSON message. No authentication. Endpoints are specified in openapi.yaml."
-}
-`,
-
-  "specs/design/components/hello-api/openapi.yaml": `openapi: 3.0.3
-info:
-  title: Hello API
-  version: 1.0.0
-  description: A simple API that responds with "Hello, World!" when called.
-
-servers:
-  - url: /
-    description: Default server
-
-paths:
-  /hello:
-    get:
-      summary: Get hello world message
-      description: Returns a simple "Hello, World!" message in JSON format.
-      operationId: getHello
-      responses:
-        '200':
-          description: Successful response with hello message
-          content:
-            application/json:
-              schema:
-                type: object
-                required:
-                  - message
-                properties:
-                  message:
-                    type: string
-                    example: "Hello, World!"
-
-  /health:
-    get:
-      summary: Health check endpoint
-      description: Returns the health status of the service.
-      operationId: getHealth
-      responses:
-        '200':
-          description: Service is healthy
-`,
-};
-
 /** Build the user prompt: the current bundle inlined + the mutation instruction. */
 export function buildPrompt(files: Record<string, string>, instruction: string): string {
   const inlined = Object.entries(files)
