+++ services/agents/test/toolset.test.ts
@@ -25,11 +25,11 @@
 import { test } from "node:test";
 import assert from "node:assert/strict";
 import { FileBundle } from "@aep/agent-stream";
-import { buildFileTools } from "./files.js";
-import { buildTaskPlanTools } from "./task-plan.js";
-import { TaskPlan } from "../task-plan-accumulator.js";
-import { instructions, buildInstructions, taskPlanInstructions, buildTaskPlanInstructions } from "../prompt.js";
-import { testSkillSource } from "../../../testing/skill-source.js";
+import { buildFileTools } from "../src/agents/main/tools/files.js";
+import { buildTaskPlanTools } from "../src/agents/main/tools/task-plan.js";
+import { TaskPlan } from "../src/agents/main/task-plan-accumulator.js";
+import { instructions, buildInstructions, taskPlanInstructions, buildTaskPlanInstructions } from "../src/agents/main/prompt.js";
+import { testSkillSource } from "./skill-source.js";
 
 const SKILLS = testSkillSource([{ name: "task-planning", description: "plan tasks", content: "one task per component" }]);
 const bundle = () => new FileBundle({});
