+++ services/agents/tsconfig.json
@@ -4,5 +4,5 @@
     "outDir": "dist",
     "noEmit": false
   },
-  "include": ["src"]
+  "include": ["src", "test"]
 }
